refactor(models): rename file names to kebab casing and make them lowercased.
