+++ components/forms/ControlBarForm.tsx
@@ -1,7 +1,7 @@
 import { cloneDeep } from "lodash";
 import { useEffect } from "react";
 import { useForm } from "react-hook-form";
-import { ConfiguratorOptions, ExtendedFluidPlayerOptions } from "@/models/ConfiguratorOptions";
+import { ConfiguratorOptions, ExtendedFluidPlayerOptions } from "@/models/configurator-options";
 import { CheckboxInput, FormField, NumberInput, TextInput } from "@/components/fields";
 import { SubmitButton } from "@/components/SubmitButton";
