+++ components/forms/MetadataForm.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useForm } from "react-hook-form";
-import { ConfiguratorOptions } from "@/models/ConfiguratorOptions";
+import { ConfiguratorOptions } from "@/models/configurator-options";
 import { FormField, TextInput } from "@/components/fields";
 import { SubmitButton } from "@/components/SubmitButton";
