+++ models/index.ts
@@ -1,2 +1,2 @@
-export * from './ConfiguratorOptions';
-export * from './FluidPlayerConfiguration';
+export * from './configurator-options';
+export * from './fluid-player-configuration';
